ocr/requirements.txt
torch
torchvision
transformers
Pillow
flask
ultralytics
opencv-python-headless